crs/src/crs-patch/smith/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.9.5
    # via
    #   datasets
    #   fsspec
    #   litellm
    #   together
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anthropic==0.30.0
    # via -r requirements.in
anyio==4.4.0
    # via
    #   anthropic
    #   httpx
    #   openai
astroid==3.2.2
    # via pylint
async-timeout==4.0.3
    # via aiohttp
attrs==23.2.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
backoff==2.2.1
    # via -r requirements.in
beautifulsoup4==4.12.3
    # via
    #   -r requirements.in
    #   gdown
    #   swebench
certifi==2024.6.2
    # via
    #   httpcore
    #   httpx
    #   requests
cfgv==3.4.0
    # via pre-commit
chardet==5.2.0
    # via swebench
charset-normalizer==3.3.2
    # via requests
clang==14.0
    # via -r requirements.in
click==8.1.7
    # via
    #   litellm
    #   together
    #   typer
cloudpickle==3.0.0
    # via gymnasium
coloredlogs==15.0.1
    # via -r requirements.in
config==0.5.1
    # via -r requirements.in
datasets==2.20.0
    # via
    #   -r requirements.in
    #   swebench
diff-match-patch==20230430
    # via -r requirements.in
dill==0.3.8
    # via
    #   datasets
    #   multiprocess
    #   pylint
diskcache==5.6.3
    # via -r requirements.in
distlib==0.3.8
    # via virtualenv
distro==1.9.0
    # via
    #   anthropic
    #   openai
docker==7.1.0
    # via
    #   -r requirements.in
    #   swebench
docstring-parser==0.16
    # via simple-parsing
eval-type-backport==0.2.0
    # via together
exceptiongroup==1.2.1
    # via
    #   anyio
    #   pytest
execnet==2.1.1
    # via pytest-xdist
farama-notifications==0.0.4
    # via gymnasium
fastcore==1.5.48
    # via ghapi
filelock==3.15.4
    # via
    #   datasets
    #   gdown
    #   huggingface-hub
    #   together
    #   virtualenv
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2024.5.0
    # via
    #   datasets
    #   huggingface-hub
gdown==5.2.0
    # via -r requirements.in
ghapi==1.0.5
    # via
    #   -r requirements.in
    #   swebench
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via swebench
greenlet==3.0.3
    # via playwright
grep-ast==0.3.2
    # via -r requirements.in
gymnasium==0.29.1
    # via -r requirements.in
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via
    #   anthropic
    #   openai
huggingface-hub==0.23.4
    # via
    #   datasets
    #   tokenizers
humanfriendly==10.0
    # via coloredlogs
identify==2.5.36
    # via pre-commit
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.0.0
    # via litellm
iniconfig==2.0.0
    # via pytest
isort==5.13.2
    # via pylint
jinja2==3.1.4
    # via litellm
jiter==0.5.0
    # via anthropic
jsonschema==4.22.0
    # via -r requirements.in
jsonschema-specifications==2023.12.1
    # via jsonschema
litellm==1.37.12
    # via -r requirements.in
lxml==5.2.2
    # via -r requirements.in
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mccabe==0.7.0
    # via pylint
mdurl==0.1.2
    # via markdown-it-py
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via datasets
networkx==3.3
    # via -r requirements.in
nodeenv==1.9.1
    # via pre-commit
numpy==2.0.0
    # via
    #   datasets
    #   gymnasium
    #   pandas
    #   pyarrow
    #   together
openai==1.23.1
    # via
    #   -r requirements.in
    #   litellm
packaging==24.1
    # via
    #   datasets
    #   fastcore
    #   ghapi
    #   huggingface-hub
    #   pytest
pandas==2.2.2
    # via
    #   -r requirements.in
    #   datasets
pathspec==0.12.1
    # via
    #   -r requirements.in
    #   grep-ast
pillow==10.3.0
    # via together
platformdirs==4.2.2
    # via
    #   pylint
    #   virtualenv
playwright==1.44.0
    # via -r requirements.in
pluggy==1.5.0
    # via pytest
pre-commit==3.7.1
    # via swebench
prompt-toolkit==3.0.47
    # via -r requirements.in
pyarrow==16.1.0
    # via
    #   datasets
    #   together
pyarrow-hotfix==0.6
    # via datasets
pydantic==2.7.4
    # via
    #   anthropic
    #   openai
    #   together
pydantic-core==2.18.4
    # via pydantic
pyee==11.1.0
    # via playwright
pygments==2.18.0
    # via rich
pylint==3.2.5
    # via -r requirements.in
pypandoc==1.13
    # via -r requirements.in
pysocks==1.7.1
    # via requests
pytest==8.2.2
    # via
    #   -r requirements.in
    #   pytest-xdist
pytest-xdist==3.6.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   pandas
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   litellm
    #   swebench
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   datasets
    #   huggingface-hub
    #   pre-commit
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.5.15
    # via tiktoken
requests[socks]==2.32.3
    # via
    #   datasets
    #   docker
    #   gdown
    #   huggingface-hub
    #   litellm
    #   swebench
    #   tiktoken
    #   together
rich==13.7.1
    # via
    #   -r requirements.in
    #   rich-argparse
    #   swebench
    #   typer
rich-argparse==1.5.2
    # via -r requirements.in
rpds-py==0.18.1
    # via
    #   jsonschema
    #   referencing
shellingham==1.5.4
    # via typer
simple-parsing==0.1.5
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   httpx
    #   openai
soupsieve==2.5
    # via beautifulsoup4
swebench==2.0.2
    # via -r requirements.in
tabulate==0.9.0
    # via together
tenacity==8.4.2
    # via -r requirements.in
tiktoken==0.7.0
    # via litellm
together==1.2.1
    # via -r requirements.in
tokenizers==0.19.1
    # via
    #   anthropic
    #   litellm
toml==0.10.2
    # via -r requirements.in
tomli==2.0.1
    # via
    #   pylint
    #   pytest
tomlkit==0.12.5
    # via pylint
tqdm==4.66.4
    # via
    #   datasets
    #   gdown
    #   huggingface-hub
    #   openai
    #   swebench
    #   together
tree-sitter==0.22.3
    # via tree-sitter-languages
tree-sitter-languages==1.10.2
    # via grep-ast
typer==0.12.3
    # via together
typing-extensions==4.12.2
    # via
    #   anthropic
    #   anyio
    #   astroid
    #   gymnasium
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   pyee
    #   simple-parsing
    #   typer
tzdata==2024.1
    # via pandas
unidiff==0.7.5
    # via -r requirements.in
urllib3==2.2.2
    # via
    #   docker
    #   requests
virtualenv==20.26.3
    # via pre-commit
wcwidth==0.2.13
    # via prompt-toolkit
xxhash==3.4.1
    # via datasets
yarl==1.9.4
    # via aiohttp
zipp==3.19.2
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
